+++ models/evc.py
@@ -1,6 +1,6 @@
 """Classes used in the main application."""  # pylint: disable=too-many-lines
 import traceback
-from collections import OrderedDict
+from collections import OrderedDict, defaultdict
 from copy import deepcopy
 from datetime import datetime
 from operator import eq, ne
@@ -22,7 +22,8 @@ from kytos.core.link import Link
 from kytos.core.retry import before_sleep
 from kytos.core.tag_ranges import range_difference
 from napps.kytos.mef_eline import controllers, settings
-from napps.kytos.mef_eline.exceptions import (DuplicatedNoTagUNI,
+from napps.kytos.mef_eline.exceptions import (ActivationError,
+                                              DuplicatedNoTagUNI,
                                               EVCPathNotInstalled,
                                               FlowModException, InvalidPath)
 from napps.kytos.mef_eline.utils import (check_disabled_component,
@@ -674,39 +675,35 @@ class EVCDeploy(EVCBase):
         """
         if not self.failover_path:
             return
-        switches, cookie, excluded = OrderedDict(), self.get_cookie(), set()
-        links = set()
+        switches, cookie, excluded = set(), self.get_cookie(), set()
         if exclude_uni_switches:
             excluded.add(self.uni_a.interface.switch.id)
             excluded.add(self.uni_z.interface.switch.id)
         for link in self.failover_path:
             if link.endpoint_a.switch.id not in excluded:
-                switches[link.endpoint_a.switch.id] = link.endpoint_a.switch
-                links.add(link)
+                switches.add(link.endpoint_a.switch.id)
             if link.endpoint_b.switch.id not in excluded:
-                switches[link.endpoint_b.switch.id] = link.endpoint_b.switch
-                links.add(link)
-        for switch in switches.values():
-            try:
-                self._send_flow_mods(
-                    switch.id,
-                    [
-                        {
-                            "cookie": cookie,
-                            "cookie_mask": int(0xffffffffffffffff),
-                            "owner": "mef_eline",
-                        }
-                    ],
-                    "delete",
-                    force=force,
-                )
-            except FlowModException as err:
-                log.error(f"Error deleting path in {switch.id}, {err}")
-                break
+                switches.add(link.endpoint_b.switch.id)
+        flow_mods = {
+            "switches": list(switches),
+            "flows": [{
+                "cookie": cookie,
+                "cookie_mask": int(0xffffffffffffffff),
+                "owner": "mef_eline",
+            }]
+        }
+        try:
+            self._send_flow_mods(
+                flow_mods,
+                "delete",
+                force=force,
+            )
+        except FlowModException as err:
+            log.error(f"Error deleting {self} failover_path flows, {err}")
         try:
             self.failover_path.make_vlans_available(self._controller)
         except KytosTagError as err:
-            log.error(f"Error when removing failover flows: {err}")
+            log.error(f"Error removing {self} failover_path: {err}")
         self.failover_path = Path([])
         if sync:
             self.sync()
@@ -732,26 +729,28 @@ class EVCDeploy(EVCBase):
 
         current_path = self.current_path
         for link in current_path:
-            switches.add(link.endpoint_a.switch)
-            switches.add(link.endpoint_b.switch)
-        switches.add(self.uni_a.interface.switch)
-        switches.add(self.uni_z.interface.switch)
-        match = {
-            "cookie": self.get_cookie(),
-            "cookie_mask": int(0xffffffffffffffff),
-            "owner": "mef_eline",
+            switches.add(link.endpoint_a.switch.id)
+            switches.add(link.endpoint_b.switch.id)
+        switches.add(self.uni_a.interface.switch.id)
+        switches.add(self.uni_z.interface.switch.id)
+        flow_mods = {
+            "switches": list(switches),
+            "flows": [{
+                "cookie": self.get_cookie(),
+                "cookie_mask": int(0xffffffffffffffff),
+                "owner": "mef_eline",
+            }]
         }
 
-        for switch in switches:
-            try:
-                self._send_flow_mods(switch.id, [match], 'delete', force=force)
-            except FlowModException as err:
-                log.error(f"Error deleting current_path in {switch.id}, {err}")
-                break
+        try:
+            self._send_flow_mods(flow_mods, "delete", force=force)
+        except FlowModException as err:
+            log.error(f"Error deleting {self} current_path flows, {err}")
+
         try:
             current_path.make_vlans_available(self._controller)
         except KytosTagError as err:
-            log.error(f"Error when removing current path flows: {err}")
+            log.error(f"Error removing {self} current_path: {err}")
             return old_path_dict
         self.current_path = Path([])
         self.deactivate()
@@ -763,7 +762,8 @@ class EVCDeploy(EVCBase):
         self, path=None, force=True
     ) -> dict[str, list[dict]]:
         """Remove all flows from path, and return the removed flows."""
-        dpid_flows_match: dict[str, list[dict]] = {}
+        dpid_flows_match: dict[str, dict] = defaultdict(lambda: {"flows": []})
+        out_flows: dict[str, list[dict]] = defaultdict(list)
 
         if not path:
             return dpid_flows_match
@@ -777,14 +777,15 @@ class EVCDeploy(EVCBase):
             nni_flows = {}
 
         for dpid, flows in nni_flows.items():
-            dpid_flows_match.setdefault(dpid, [])
             for flow in flows:
-                dpid_flows_match[dpid].append({
+                flow_mod = {
                     "cookie": flow["cookie"],
                     "match": flow["match"],
                     "owner": "mef_eline",
                     "cookie_mask": int(0xffffffffffffffff)
-                })
+                }
+                dpid_flows_match[dpid]["flows"].append(flow_mod)
+                out_flows[dpid].append(flow_mod)
 
         try:
             uni_flows = self._prepare_uni_flows(path, skip_in=True)
@@ -795,27 +796,31 @@ class EVCDeploy(EVCBase):
             uni_flows = {}
 
         for dpid, flows in uni_flows.items():
-            dpid_flows_match.setdefault(dpid, [])
             for flow in flows:
-                dpid_flows_match[dpid].append({
+                flow_mod = {
                     "cookie": flow["cookie"],
                     "match": flow["match"],
                     "owner": "mef_eline",
                     "cookie_mask": int(0xffffffffffffffff)
-                })
+                }
+                dpid_flows_match[dpid]["flows"].append(flow_mod)
+                out_flows[dpid].append(flow_mod)
 
-        for dpid, flows in dpid_flows_match.items():
-            try:
-                self._send_flow_mods(dpid, flows, 'delete', force=force)
-            except FlowModException as err:
-                log.error(f"Error deleting path in {dpid}, {err}")
-                break
+        try:
+            self._send_flow_mods(
+                dpid_flows_match, 'delete', force=force, by_switch=True
+            )
+        except FlowModException as err:
+            log.error(
+                f"Error deleting {self} path flows, path:{path}, error={err}"
+            )
+
         try:
             path.make_vlans_available(self._controller)
         except KytosTagError as err:
-            log.error(f"Error when removing path flows: {err}")
+            log.error(f"Error removing {self} path: {err}")
 
-        return dpid_flows_match
+        return out_flows
 
     @staticmethod
     def links_zipped(path=None):
@@ -839,6 +844,61 @@ class EVCDeploy(EVCBase):
             return True
 
         return False
+
+    @staticmethod
+    def is_uni_interface_active(
+        *interfaces: Interface
+    ) -> tuple[bool, dict]:
+        """Whether UNIs are active and their status & status_reason."""
+        active = True
+        bad_interfaces = [
+            interface
+            for interface in interfaces
+            if interface.status != EntityStatus.UP
+        ]
+        if bad_interfaces:
+            active = False
+            interfaces = bad_interfaces
+        return active, {
+            interface.id: {
+                'status': interface.status.value,
+                'status_reason': interface.status_reason,
+            }
+            for interface in interfaces
+        }
+
+    def try_to_activate(self) -> bool:
+        """Try to activate the EVC."""
+        if self.is_intra_switch():
+            return self._try_to_activate_intra_evc()
+        return self._try_to_activate_inter_evc()
+
+    def _try_to_activate_intra_evc(self) -> bool:
+        """Try to activate intra EVC."""
+        intf_a, intf_z = self.uni_a.interface, self.uni_z.interface
+        is_active, reason = self.is_uni_interface_active(intf_a, intf_z)
+        if not is_active:
+            raise ActivationError(
+                f"Won't be able to activate {self} due to UNIs: {reason}"
+            )
+        self.activate()
+        return True
+
+    def _try_to_activate_inter_evc(self) -> bool:
+        """Try to activate inter EVC."""
+        intf_a, intf_z = self.uni_a.interface, self.uni_z.interface
+        is_active, reason = self.is_uni_interface_active(intf_a, intf_z)
+        if not is_active:
+            raise ActivationError(
+                f"Won't be able to activate {self} due to UNIs: {reason}"
+            )
+        if self.current_path.status != EntityStatus.UP:
+            raise ActivationError(
+                f"Won't be able to activate {self} due to current_path "
+                f"status {self.current_path.status}"
+            )
+        self.activate()
+        return True
 
     # pylint: disable=too-many-branches, too-many-statements
     def deploy_to_path(self, path=None, old_path_dict: dict = None):
@@ -881,8 +941,7 @@ class EVCDeploy(EVCBase):
 
         try:
             if use_path:
-                self._install_nni_flows(use_path)
-                self._install_uni_flows(use_path)
+                self._install_flows(use_path)
             elif self.is_intra_switch():
                 use_path = Path()
                 self._install_direct_uni_flows()
@@ -900,10 +959,15 @@ class EVCDeploy(EVCBase):
             )
             self.remove_current_flows(use_path, sync=True)
             return False
-        self.activate()
+
         self.current_path = use_path
+        msg = f"{self} was deployed."
+        try:
+            self.try_to_activate()
+        except ActivationError as exc:
+            msg = f"{msg} {str(exc)}"
         self.sync()
-        log.info(f"{self} was deployed.")
+        log.info(msg)
         return True
 
     def try_setup_failover_path(self, wait=settings.DEPLOY_EVCS_INTERVAL):
@@ -958,10 +1022,9 @@ class EVCDeploy(EVCBase):
 
         try:
             if use_path:
-                _nni_flows = self._install_nni_flows(use_path)
-                out_new_flows = merge_flow_dicts(out_new_flows, _nni_flows)
-                _uni_flows = self._install_uni_flows(use_path, skip_in=True)
-                out_new_flows = merge_flow_dicts(out_new_flows, _uni_flows)
+                out_new_flows = self._install_flows(
+                    use_path, skip_in=True
+                )
         except EVCPathNotInstalled as err:
             reason = "Error deploying failover path"
             log.error(
@@ -1098,10 +1161,11 @@ class EVCDeploy(EVCBase):
         same switch.
         """
         (dpid, flows) = self._prepare_direct_uni_flows()
+        flow_mods = {"switches": [dpid], "flows": flows}
         try:
-            self._send_flow_mods(dpid, flows)
+            self._send_flow_mods(flow_mods, "install")
         except FlowModException as err:
-            raise EVCPathNotInstalled(f"In {dpid}, {err}") from err
+            raise EVCPathNotInstalled(str(err)) from err
 
     def _prepare_nni_flows(self, path=None):
         """Prepare NNI flows."""
@@ -1141,17 +1205,27 @@ class EVCDeploy(EVCBase):
             nni_flows[in_endpoint.switch.id] = flows
         return nni_flows
 
-    def _install_nni_flows(
-        self, path=None
+    def _install_flows(
+        self, path=None, skip_in=False, skip_out=False
     ) -> dict[str, list[dict]]:
-        """Install NNI flows."""
-        nni_flows = self._prepare_nni_flows(path)
+        """Install uni and nni flows"""
+        flows_by_switch = defaultdict(lambda: {"flows": []})
+        new_flows = defaultdict(list)
+        for dpid, flows in self._prepare_nni_flows(path).items():
+            flows_by_switch[dpid]["flows"].extend(flows)
+            new_flows[dpid].extend(flows)
+        for dpid, flows in self._prepare_uni_flows(
+            path, skip_in, skip_out
+        ).items():
+            flows_by_switch[dpid]["flows"].extend(flows)
+            new_flows[dpid].extend(flows)
+
         try:
-            for dpid, flows in nni_flows.items():
-                self._send_flow_mods(dpid, flows)
+            self._send_flow_mods(flows_by_switch, "install", by_switch=True)
         except FlowModException as err:
-            raise EVCPathNotInstalled(f"In {dpid}, {err}") from err
-        return nni_flows
+            raise EVCPathNotInstalled(str(err)) from err
+
+        return new_flows
 
     @staticmethod
     def _get_value_from_uni_tag(uni: UNI):
@@ -1268,19 +1342,6 @@ class EVCDeploy(EVCBase):
 
         return uni_flows
 
-    def _install_uni_flows(
-        self, path=None, skip_in=False, skip_out=False
-    ) -> dict[str, list[dict]]:
-        """Install UNI flows."""
-        uni_flows = self._prepare_uni_flows(path, skip_in, skip_out)
-        try:
-            for (dpid, flows) in uni_flows.items():
-                self._send_flow_mods(dpid, flows)
-        except FlowModException as err:
-            raise EVCPathNotInstalled(f"In {dpid}, {err}") from err
-
-        return uni_flows
-
     @staticmethod
     @retry(
         stop=stop_after_attempt(3),
@@ -1289,21 +1350,33 @@ class EVCDeploy(EVCBase):
         before_sleep=before_sleep,
         reraise=True,
     )
-    def _send_flow_mods(dpid, flow_mods, command='flows', force=False):
+    def _send_flow_mods(
+        data_content: dict,
+        command="install",
+        force=False,
+        by_switch=False
+    ):
         """Send a flow_mod list to a specific switch.
 
         Args:
             dpid(str): The target of flows (i.e. Switch.id).
             flow_mods(dict): Python dictionary with flow_mods.
             command(str): By default is 'flows'. To remove a flow is 'remove'.
-            force(bool): True to send via consistency check in case of errors
-
+            force(bool): True to send via consistency check in case of errors.
+            by_switch(bool): True to send to 'flows_by_switch' request instead.
         """
-        endpoint = f"{settings.MANAGER_URL}/{command}/{dpid}"
-
-        data = {"flows": flow_mods, "force": force}
+        if by_switch:
+            endpoint = f"{settings.MANAGER_URL}/flows_by_switch/?force={force}"
+        else:
+            endpoint = f"{settings.MANAGER_URL}/flows"
+            data_content["force"] = force
         try:
-            res = httpx.post(endpoint, json=data, timeout=30)
+            if command == "install":
+                res = httpx.post(endpoint, json=data_content, timeout=30)
+            elif command == "delete":
+                res = httpx.request(
+                    "DELETE", endpoint, json=data_content, timeout=30
+                )
         except httpx.RequestError as err:
             raise FlowModException(str(err)) from err
         if res.is_server_error or res.status_code >= 400:
@@ -1758,27 +1831,26 @@ class LinkProtection(EVCDeploy):
         active, _ = self.is_uni_interface_active(interface_a, interface_z)
         return active
 
-    @staticmethod
-    def is_uni_interface_active(
-        *interfaces: Interface
-    ) -> tuple[bool, dict]:
-        """Determine whether a UNI should be active"""
-        active = True
-        bad_interfaces = [
-            interface
-            for interface in interfaces
-            if interface.status != EntityStatus.UP
-        ]
-        if bad_interfaces:
-            active = False
-            interfaces = bad_interfaces
-        return active, {
-            interface.id: {
-                'status': interface.status.value,
-                'status_reason': interface.status_reason,
-            }
-            for interface in interfaces
-        }
+    def try_to_handle_uni_as_link_up(self, interface: Interface) -> bool:
+        """Try to handle UNI as link_up to trigger deployment."""
+        if (
+            self.current_path.status != EntityStatus.UP
+            and not self.is_intra_switch()
+        ):
+            succeeded = self.handle_link_up(interface)
+            if succeeded:
+                msg = (
+                    f"Activated {self} due to successful "
+                    f"deployment triggered by {interface}"
+                )
+            else:
+                msg = (
+                    f"Couldn't activate {self} due to unsuccessful "
+                    f"deployment triggered by {interface}"
+                )
+            log.info(msg)
+            return True
+        return False
 
     def handle_interface_link_up(self, interface: Interface):
         """
@@ -1796,6 +1868,9 @@ class LinkProtection(EVCDeploy):
         ]
         if down_interfaces:
             return
+        if self.try_to_handle_uni_as_link_up(interface):
+            return
+
         interface_dicts = {
             interface.id: {
                 'status': interface.status.value,
@@ -1803,14 +1878,19 @@ class LinkProtection(EVCDeploy):
             }
             for interface in interfaces
         }
-        self.activate()
-        log.info(
-            f"Activating {self}. Interfaces: "
-            f"{interface_dicts}."
-        )
-        emit_event(self._controller, "uni_active_updated",
-                   content=map_evc_event_content(self))
-        self.sync()
+        try:
+            self.try_to_activate()
+            log.info(
+                f"Activating {self}. Interfaces: "
+                f"{interface_dicts}."
+            )
+            emit_event(self._controller, "uni_active_updated",
+                       content=map_evc_event_content(self))
+            self.sync()
+        except ActivationError as exc:
+            # On this ctx, no ActivationError isn't expected since the
+            # activation pre-requisites states were checked, so handled as err
+            log.error(f"ActivationError: {str(exc)} when handling {interface}")
 
     def handle_interface_link_down(self, interface):
         """